Refactor module interop as a folder into a file interface
