--- ams/interop/__init__.py
@@ -1,18 +1,0 @@
-"""
-Interopability package between AMS and other software.
-
-To install dependencies, do:
-
-.. code:: bash
-
-    pip install ams[interop]
-
-To install dependencies for *development*, in the AMS source code folder, do:
-
-.. code:: bash
-
-    pip install -e .[interop]
-
-"""
-
-from ams.interop import andes  # NOQA
